refactor: CartViewModel 의존성 제거
- recentItemRepository 의존성 제거
- productRepository 의존성 제거
- cart ui 수정

+++ app/src/main/java/woowacourse/shopping/ui/cart/CartScreen.kt
@@ -112,7 +112,7 @@ fun CartScreen(
 
 @Composable
 private fun CartContent(
-    totalCartSize: Int,
+    totalCartSize: Long,
     page: Int,
     onNextPage: () -> Unit,
     onPreviousPage: () -> Unit,
